requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.1.15
django-extensions==2.2.8
django-mysql==3.3.0
djangorestframework==3.11.0
idna==2.8
mysqlclient==1.4.6
pytz==2019.3
requests==2.22.0
six==1.14.0
stripe==2.42.0
urllib3==1.25.8
